// src/types/globalTypes.ts
export enum PermissionEnum {
    HANDLE_PAYMENTS = "HANDLE_PAYMENTS",
    MANAGE_APPS = "MANAGE_APPS",
    MANAGE_CHANNELS = "MANAGE_CHANNELS",
    MANAGE_CHECKOUTS = "MANAGE_CHECKOUTS",
    MANAGE_DISCOUNTS = "MANAGE_DISCOUNTS",
    MANAGE_GIFT_CARD = "MANAGE_GIFT_CARD",
    MANAGE_MENUS = "MANAGE_MENUS",
    MANAGE_ORDERS = "MANAGE_ORDERS",
    MANAGE_PAGES = "MANAGE_PAGES",
    MANAGE_PAGE_TYPES_AND_ATTRIBUTES = "MANAGE_PAGE_TYPES_AND_ATTRIBUTES",
    MANAGE_PLUGINS = "MANAGE_PLUGINS",
    MANAGE_PRODUCTS = "MANAGE_PRODUCTS",
    MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES = "MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES",
    MANAGE_SETTINGS = "MANAGE_SETTINGS",
    MANAGE_SHIPPING = "MANAGE_SHIPPING",
    MANAGE_STAFF = "MANAGE_STAFF",
    MANAGE_TRANSLATIONS = "MANAGE_TRANSLATIONS",
    MANAGE_USERS = "MANAGE_USERS",
}